requirements.txt
cffi==1.11.5
Django==1.11.10
django-filter==1.1.0
djangorestframework==3.7.7
factory-boy==2.10.0
google-api-python-client==1.6.5
gunicorn==19.7.1
httplib2==0.10.3
oauth2client==4.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-ldap==2.5.2
requests==2.18.4
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
psycopg2-binary
